src/requirements.txt
auto_mix_prep==0.2.0
beautifulsoup4==4.11.1
cohere==2.6.1
psycopg==3.1.3
python-dotenv==0.21.0
python_bcrypt==0.3.2
requests==2.28.1
